Separate content iterator for test bench
