+++ src/main/kotlin/com/mituuz/fuzzier/Fuzzier.kt
@@ -26,6 +26,7 @@ import com.intellij.openapi.util.DimensionService
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.openapi.vfs.VirtualFileManager
 import com.intellij.openapi.wm.WindowManager
+import com.mituuz.fuzzier.StringEvaluator.FuzzyMatchContainer
 import com.mituuz.fuzzier.settings.FuzzierSettingsService
 import org.apache.commons.lang3.StringUtils
 import java.awt.event.*
@@ -303,8 +304,6 @@ class Fuzzier : AnAction() {
 
         return score.toInt()
     }
-
-    data class FuzzyMatchContainer(val score: Int, val string: String)
 
     private fun openFile(project: Project, virtualFile: VirtualFile) {
         val fileEditorManager = FileEditorManager.getInstance(project)
